question_answering_app_LLM/requirements.txt
langchain
Openai
Streamlit
langchain_community
python-dotenv